requirements.txt
streamlit==1.41.1
langchain-community==0.3.13
langchain-openai==0.2.14
faiss-cpu==1.9.0.post1
replicate
google-cloud-storage
google-cloud-secret-manager